+++ Source/Csla.AspNetCore/Blazor/ApplicationContextManagerBlazor.cs
@@ -7,6 +7,7 @@
 // <summary>Application context manager that uses HttpContextAccessor</summary>
 //-----------------------------------------------------------------------
 using Csla.Core;
+using Csla.State;
 using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.AspNetCore.Http;
 using System;
@@ -23,6 +24,23 @@ namespace Csla.AspNetCore.Blazor
   /// </summary>
   public class ApplicationContextManagerBlazor : IContextManager, IDisposable
   {
+    /// <summary>
+    /// Creates an instance of the object, initializing it
+    /// with the required IServiceProvider.
+    /// </summary>
+    /// <param name="hca">IHttpContextAccessor service</param>
+    /// <param name="authenticationStateProvider">AuthenticationStateProvider service</param>
+    /// <param name="activeCircuitState"></param>
+    public ApplicationContextManagerBlazor(IHttpContextAccessor hca, AuthenticationStateProvider authenticationStateProvider, ActiveCircuitState activeCircuitState)
+    {
+      HttpContext = hca.HttpContext;
+      AuthenticationStateProvider = authenticationStateProvider;
+      ActiveCircuitState = activeCircuitState;
+      CurrentPrincipal = UnauthenticatedPrincipal;
+      AuthenticationStateProvider.AuthenticationStateChanged += AuthenticationStateProvider_AuthenticationStateChanged;
+      InitializeUser();
+    }
+
     private ContextDictionary LocalContext { get; set; }
     private ContextDictionary ClientContext { get; set; }
     private IPrincipal CurrentPrincipal { get; set; }
@@ -48,23 +66,6 @@ namespace Csla.AspNetCore.Blazor
     /// Gets or sets a reference to the current HttpContext.
     /// </summary>
     protected HttpContext HttpContext { get; set; }
-
-    /// <summary>
-    /// Creates an instance of the object, initializing it
-    /// with the required IServiceProvider.
-    /// </summary>
-    /// <param name="hca">IHttpContextAccessor service</param>
-    /// <param name="authenticationStateProvider">AuthenticationStateProvider service</param>
-    /// <param name="activeCircuitState"></param>
-    public ApplicationContextManagerBlazor(IHttpContextAccessor hca, AuthenticationStateProvider authenticationStateProvider, ActiveCircuitState activeCircuitState)
-    {
-      HttpContext = hca.HttpContext;
-      AuthenticationStateProvider = authenticationStateProvider;
-      ActiveCircuitState = activeCircuitState;
-      CurrentPrincipal = UnauthenticatedPrincipal;
-      AuthenticationStateProvider.AuthenticationStateChanged += AuthenticationStateProvider_AuthenticationStateChanged;
-      InitializeUser();
-    }
 
     private void InitializeUser()
     {
@@ -186,8 +187,17 @@ namespace Csla.AspNetCore.Blazor
     /// </summary>
     public ContextDictionary GetLocalContext()
     {
-      if (LocalContext == null)
-        LocalContext = new ContextDictionary();
+      if (LocalContext is null)
+      {
+        var sessionManager = ApplicationContext.GetRequiredService<ISessionManager>();
+        var session = sessionManager.GetSession();
+        session.TryGetValue("localContext", out var result);
+        if (result is ContextDictionary context)
+          LocalContext = context;
+        else
+          LocalContext = [];
+        SetLocalContext(LocalContext);
+      }
       return LocalContext;
     }
 
@@ -198,6 +208,9 @@ namespace Csla.AspNetCore.Blazor
     public void SetLocalContext(ContextDictionary localContext)
     {
       LocalContext = localContext;
+      var sessionManager = ApplicationContext.GetRequiredService<ISessionManager>();
+      var session = sessionManager.GetSession();
+      session["localContext"] = localContext;
     }
 
     /// <summary>
@@ -206,8 +219,17 @@ namespace Csla.AspNetCore.Blazor
     /// <param name="executionLocation"></param>
     public ContextDictionary GetClientContext(ApplicationContext.ExecutionLocations executionLocation)
     {
-      if (ClientContext == null)
-        ClientContext = new ContextDictionary();
+      if (ClientContext is null)
+      {
+        var sessionManager = ApplicationContext.GetRequiredService<ISessionManager>();
+        var session = sessionManager.GetSession();
+        session.TryGetValue("clientContext", out var result);
+        if (result is ContextDictionary context)
+          ClientContext = context;
+        else
+          ClientContext = [];
+        SetClientContext(ClientContext, ApplicationContext.ExecutionLocation);
+      }
       return ClientContext;
     }
 
@@ -219,6 +241,9 @@ namespace Csla.AspNetCore.Blazor
     public void SetClientContext(ContextDictionary clientContext, ApplicationContext.ExecutionLocations executionLocation)
     {
       ClientContext = clientContext;
+      var sessionManager = ApplicationContext.GetRequiredService<ISessionManager>();
+      var session = sessionManager.GetSession();
+      session["clientContext"] = clientContext;
     }
 
     /// <summary>